ponytail: extract shared VerifyEmailForm, trim dead code
- New components/VerifyEmailForm.tsx shared between user and organiser
  portals, saving ~355 lines of duplication
- Removed unused sessionStorage keys in SignInModal
- Simplified submitToApi return type (Promise<void>, no caller checked it)
- Regenerated Prisma client (tests were failing without it)
